Add unittest for particle gun and split headers
